+++ app/src/components/ProfileDropDown/ProfileDropDown.js
@@ -19,7 +19,8 @@ class ProfileDropDown extends React.Component {
     const profile = this.props.profile;
     return (
       <div >
-        <NavDropdown style={{display: 'inline-block'}} eventKey={5} id='basic-nav-dropdown' title={profile.firstName + ' ' + profile.lastName}>
+        <NavDropdown style={{display: 'inline-block'}} eventKey={5} id='basic-nav-dropdown'
+          title={profile.firstName + ' ' + profile.lastName}>
           <LinkContainer to='/profile'>
             <MenuItem eventKey={5.1}>Edit Profile</MenuItem>
           </LinkContainer>
