+++ app/src/components/index.js
@@ -1,4 +1,4 @@
 export Navigation from './Navigation/Navigation';
-export ProfileDropdown from './Navigation/ProfileDropDown';
+export ProfileDropdown from './ProfileDropDown/ProfileDropDown';
 export Login from './Login/Login';
 export Message from './Message/Message';
